replace mamba_lm.py with generic lm.py
